Added new core.docutils submodule, which allows the configuration details to be split more easily across pages
