+++ docs/source/_toc.yaml
@@ -58,14 +58,19 @@ subtrees:
           - file: using_the_ve/variables/variables
           - file: using_the_ve/core_settings/overview
             entries:
-              - file: using_the_ve/core_configuration
+              - file: using_the_ve/core_settings/core_configuration
               - file: using_the_ve/axes
               - file: using_the_ve/core_settings/static_models
           - file: using_the_ve/model_details/overview
             entries:
+              - file: using_the_ve/model_details/abiotic/overview
+                entries:
+                  - file: using_the_ve/model_details/abiotic/abiotic_config
+                  - file: using_the_ve/model_details/abiotic/abiotic_simple_config
+                  - file: using_the_ve/model_details/abiotic/hydrology_config
+                  - file: using_the_ve/model_details/abiotic/notes_preprocessing
               - file: using_the_ve/model_details/pft_configuration
-              - file: using_the_ve/model_details/notes_preprocessing
-              - file: using_the_ve/science_model_configuration
+              - file: using_the_ve/model_details/science_model_configuration
               - file: using_the_ve/model_details/new_config
 
       - file: development/contributing
@@ -125,6 +130,8 @@ subtrees:
                     title: The utils submodule
                   - file: api/core/variables.md
                     title: The variables submodule
+                  - file: api/core/docutils.md
+                    title: The docutils submodule
           - file: api/main.md
             title: The main submodule
           - file: api/models.md
