+++ docs/source/api/core/docutils.md
@@ -1,0 +1,34 @@
+---
+jupytext:
+  cell_metadata_filter: -all
+  formats: md:myst
+  main_language: python
+  text_representation:
+    extension: .md
+    format_name: myst
+    format_version: 0.13
+    jupytext_version: 1.19.1
+kernelspec:
+  display_name: Python 3 (ipykernel)
+  language: python
+  name: python3
+language_info:
+  codemirror_mode:
+    name: ipython
+    version: 3
+  file_extension: .py
+  mimetype: text/x-python
+  name: python
+  nbconvert_exporter: python
+  pygments_lexer: ipython3
+  version: 3.11.9
+---
+
+# API documentation for the {mod}`~virtual_ecosystem.core.docutils` module
+
+```{eval-rst}
+.. automodule:: virtual_ecosystem.core.docutils
+    :autosummary:
+    :members:
+    :special-members: __init_subclass__
+```
