+++ docs/source/using_the_ve/model_details/abiotic/overview.md
@@ -1,0 +1,27 @@
+---
+jupyter:
+  jupytext:
+    cell_metadata_filter: all,-trusted
+    main_language: python
+    notebook_metadata_filter: settings,mystnb,language_info,execution
+    text_representation:
+      extension: .md
+      format_name: markdown
+      format_version: '1.3'
+      jupytext_version: 1.19.1
+---
+
+# The abiotic environment models
+
+The Virtual Ecosystem includes three different abiotic environmental models. For most
+simulations, you will want to include a model of microclimate, the choices are either a
+process based model ([`abiotic`](./abiotic_config.md)) or a less
+computational intensive model
+([`abiotic_simple`](./abiotic_simple_config.md)). You will also generally
+want to include a model of hydrological processes, for which
+[`hydrology`](./hydrology_config.md) is the only option.
+
+## Preparing the input data for the abiotic models
+
+The first thing you should do before you start preparing your input data is read [these
+important notes about climate data pre-processing](./notes_preprocessing.md).
